requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
defusedxml==0.7.1
filelock==3.9.0
Glances==3.3.1.1
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gspread==5.7.2
httplib2==0.21.0
huggingface-hub==0.12.1
idna==3.4
oauthlib==3.2.2
packaging==23.0
protobuf==4.22.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
typing_extensions==4.5.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
Flask==2.3.2  # or the version you're using
python-dotenv==1.0.0  # or any specific version you'd like to use